refactor dask
