+++ narwhals/_dask/expr.py
@@ -7,20 +7,20 @@ from typing import Literal
 from typing import NoReturn
 from typing import Sequence
 
+from narwhals._dask.expr_dt import DaskExprDateTimeNamespace
+from narwhals._dask.expr_name import DaskExprNameNamespace
+from narwhals._dask.expr_str import DaskExprStringNamespace
 from narwhals._dask.utils import add_row_index
 from narwhals._dask.utils import binary_operation_returns_scalar
 from narwhals._dask.utils import maybe_evaluate
 from narwhals._dask.utils import narwhals_to_native_dtype
 from narwhals._expression_parsing import infer_new_root_output_names
-from narwhals._pandas_like.utils import calculate_timestamp_date
-from narwhals._pandas_like.utils import calculate_timestamp_datetime
 from narwhals._pandas_like.utils import native_to_narwhals_dtype
 from narwhals.exceptions import ColumnNotFoundError
 from narwhals.exceptions import InvalidOperationError
 from narwhals.typing import CompliantExpr
 from narwhals.utils import Implementation
 from narwhals.utils import generate_temporary_column_name
-from narwhals.utils import import_dtypes_module
 
 if TYPE_CHECKING:
     try:
@@ -722,18 +722,6 @@ class DaskExpr(CompliantExpr["dx.Series"]):
             kwargs={**self._kwargs, "keys": keys},
         )
 
-    @property
-    def str(self: Self) -> DaskExprStringNamespace:
-        return DaskExprStringNamespace(self)
-
-    @property
-    def dt(self: Self) -> DaskExprDateTimeNamespace:
-        return DaskExprDateTimeNamespace(self)
-
-    @property
-    def name(self: Self) -> DaskExprNameNamespace:
-        return DaskExprNameNamespace(self)
-
     def cast(self: Self, dtype: DType | type[DType]) -> Self:
         def func(_input: Any, dtype: DType | type[DType]) -> Any:
             dtype = narwhals_to_native_dtype(dtype, self._version)
@@ -752,470 +740,14 @@ class DaskExpr(CompliantExpr["dx.Series"]):
             returns_scalar=self._returns_scalar,
         )
 
+    @property
+    def str(self: Self) -> DaskExprStringNamespace:
+        return DaskExprStringNamespace(self)
 
-class DaskExprStringNamespace:
-    def __init__(self, expr: DaskExpr) -> None:
-        self._compliant_expr = expr
+    @property
+    def dt(self: Self) -> DaskExprDateTimeNamespace:
+        return DaskExprDateTimeNamespace(self)
 
-    def len_chars(self) -> DaskExpr:
-        return self._compliant_expr._from_call(
-            lambda _input: _input.str.len(),
-            "len",
-            returns_scalar=self._compliant_expr._returns_scalar,
-        )
-
-    def replace(
-        self, pattern: str, value: str, *, literal: bool = False, n: int = 1
-    ) -> DaskExpr:
-        return self._compliant_expr._from_call(
-            lambda _input, pattern, value, literal, n: _input.str.replace(
-                pattern, value, regex=not literal, n=n
-            ),
-            "replace",
-            pattern=pattern,
-            value=value,
-            literal=literal,
-            n=n,
-            returns_scalar=self._compliant_expr._returns_scalar,
-        )
-
-    def replace_all(self, pattern: str, value: str, *, literal: bool = False) -> DaskExpr:
-        return self._compliant_expr._from_call(
-            lambda _input, pattern, value, literal: _input.str.replace(
-                pattern, value, n=-1, regex=not literal
-            ),
-            "replace",
-            pattern=pattern,
-            value=value,
-            literal=literal,
-            returns_scalar=self._compliant_expr._returns_scalar,
-        )
-
-    def strip_chars(self, characters: str | None = None) -> DaskExpr:
-        return self._compliant_expr._from_call(
-            lambda _input, characters: _input.str.strip(characters),
-            "strip",
-            characters=characters,
-            returns_scalar=self._compliant_expr._returns_scalar,
-        )
-
-    def starts_with(self, prefix: str) -> DaskExpr:
-        return self._compliant_expr._from_call(
-            lambda _input, prefix: _input.str.startswith(prefix),
-            "starts_with",
-            prefix=prefix,
-            returns_scalar=self._compliant_expr._returns_scalar,
-        )
-
-    def ends_with(self, suffix: str) -> DaskExpr:
-        return self._compliant_expr._from_call(
-            lambda _input, suffix: _input.str.endswith(suffix),
-            "ends_with",
-            suffix=suffix,
-            returns_scalar=self._compliant_expr._returns_scalar,
-        )
-
-    def contains(self, pattern: str, *, literal: bool = False) -> DaskExpr:
-        return self._compliant_expr._from_call(
-            lambda _input, pattern, literal: _input.str.contains(
-                pat=pattern, regex=not literal
-            ),
-            "contains",
-            pattern=pattern,
-            literal=literal,
-            returns_scalar=self._compliant_expr._returns_scalar,
-        )
-
-    def slice(self, offset: int, length: int | None = None) -> DaskExpr:
-        return self._compliant_expr._from_call(
-            lambda _input, offset, length: _input.str.slice(
-                start=offset, stop=offset + length if length else None
-            ),
-            "slice",
-            offset=offset,
-            length=length,
-            returns_scalar=self._compliant_expr._returns_scalar,
-        )
-
-    def to_datetime(self: Self, format: str | None) -> DaskExpr:  # noqa: A002
-        import dask.dataframe as dd
-
-        return self._compliant_expr._from_call(
-            lambda _input, format: dd.to_datetime(_input, format=format),
-            "to_datetime",
-            format=format,
-            returns_scalar=self._compliant_expr._returns_scalar,
-        )
-
-    def to_uppercase(self) -> DaskExpr:
-        return self._compliant_expr._from_call(
-            lambda _input: _input.str.upper(),
-            "to_uppercase",
-            returns_scalar=self._compliant_expr._returns_scalar,
-        )
-
-    def to_lowercase(self) -> DaskExpr:
-        return self._compliant_expr._from_call(
-            lambda _input: _input.str.lower(),
-            "to_lowercase",
-            returns_scalar=self._compliant_expr._returns_scalar,
-        )
-
-
-class DaskExprDateTimeNamespace:
-    def __init__(self, expr: DaskExpr) -> None:
-        self._compliant_expr = expr
-
-    def date(self) -> DaskExpr:
-        return self._compliant_expr._from_call(
-            lambda _input: _input.dt.date,
-            "date",
-            returns_scalar=self._compliant_expr._returns_scalar,
-        )
-
-    def year(self) -> DaskExpr:
-        return self._compliant_expr._from_call(
-            lambda _input: _input.dt.year,
-            "year",
-            returns_scalar=self._compliant_expr._returns_scalar,
-        )
-
-    def month(self) -> DaskExpr:
-        return self._compliant_expr._from_call(
-            lambda _input: _input.dt.month,
-            "month",
-            returns_scalar=self._compliant_expr._returns_scalar,
-        )
-
-    def day(self) -> DaskExpr:
-        return self._compliant_expr._from_call(
-            lambda _input: _input.dt.day,
-            "day",
-            returns_scalar=self._compliant_expr._returns_scalar,
-        )
-
-    def hour(self) -> DaskExpr:
-        return self._compliant_expr._from_call(
-            lambda _input: _input.dt.hour,
-            "hour",
-            returns_scalar=self._compliant_expr._returns_scalar,
-        )
-
-    def minute(self) -> DaskExpr:
-        return self._compliant_expr._from_call(
-            lambda _input: _input.dt.minute,
-            "minute",
-            returns_scalar=self._compliant_expr._returns_scalar,
-        )
-
-    def second(self) -> DaskExpr:
-        return self._compliant_expr._from_call(
-            lambda _input: _input.dt.second,
-            "second",
-            returns_scalar=self._compliant_expr._returns_scalar,
-        )
-
-    def millisecond(self) -> DaskExpr:
-        return self._compliant_expr._from_call(
-            lambda _input: _input.dt.microsecond // 1000,
-            "millisecond",
-            returns_scalar=self._compliant_expr._returns_scalar,
-        )
-
-    def microsecond(self) -> DaskExpr:
-        return self._compliant_expr._from_call(
-            lambda _input: _input.dt.microsecond,
-            "microsecond",
-            returns_scalar=self._compliant_expr._returns_scalar,
-        )
-
-    def nanosecond(self) -> DaskExpr:
-        return self._compliant_expr._from_call(
-            lambda _input: _input.dt.microsecond * 1000 + _input.dt.nanosecond,
-            "nanosecond",
-            returns_scalar=self._compliant_expr._returns_scalar,
-        )
-
-    def ordinal_day(self) -> DaskExpr:
-        return self._compliant_expr._from_call(
-            lambda _input: _input.dt.dayofyear,
-            "ordinal_day",
-            returns_scalar=self._compliant_expr._returns_scalar,
-        )
-
-    def weekday(self) -> DaskExpr:
-        return self._compliant_expr._from_call(
-            lambda _input: _input.dt.weekday + 1,  # Dask is 0-6
-            "weekday",
-            returns_scalar=self._compliant_expr._returns_scalar,
-        )
-
-    def to_string(self, format: str) -> DaskExpr:  # noqa: A002
-        return self._compliant_expr._from_call(
-            lambda _input, format: _input.dt.strftime(format.replace("%.f", ".%f")),
-            "strftime",
-            format=format,
-            returns_scalar=self._compliant_expr._returns_scalar,
-        )
-
-    def replace_time_zone(self, time_zone: str | None) -> DaskExpr:
-        return self._compliant_expr._from_call(
-            lambda _input, time_zone: _input.dt.tz_localize(None).dt.tz_localize(
-                time_zone
-            )
-            if time_zone is not None
-            else _input.dt.tz_localize(None),
-            "tz_localize",
-            time_zone=time_zone,
-            returns_scalar=self._compliant_expr._returns_scalar,
-        )
-
-    def convert_time_zone(self, time_zone: str) -> DaskExpr:
-        def func(s: dx.Series, time_zone: str) -> dx.Series:
-            dtype = native_to_narwhals_dtype(
-                s, self._compliant_expr._version, Implementation.DASK
-            )
-            if dtype.time_zone is None:  # type: ignore[attr-defined]
-                return s.dt.tz_localize("UTC").dt.tz_convert(time_zone)
-            else:
-                return s.dt.tz_convert(time_zone)
-
-        return self._compliant_expr._from_call(
-            func,
-            "tz_convert",
-            time_zone=time_zone,
-            returns_scalar=self._compliant_expr._returns_scalar,
-        )
-
-    def timestamp(self, time_unit: Literal["ns", "us", "ms"] = "us") -> DaskExpr:
-        def func(s: dx.Series, time_unit: Literal["ns", "us", "ms"] = "us") -> dx.Series:
-            dtype = native_to_narwhals_dtype(
-                s, self._compliant_expr._version, Implementation.DASK
-            )
-            is_pyarrow_dtype = "pyarrow" in str(dtype)
-            mask_na = s.isna()
-            dtypes = import_dtypes_module(self._compliant_expr._version)
-            if dtype == dtypes.Date:
-                # Date is only supported in pandas dtypes if pyarrow-backed
-                s_cast = s.astype("Int32[pyarrow]")
-                result = calculate_timestamp_date(s_cast, time_unit)
-            elif dtype == dtypes.Datetime:
-                original_time_unit = dtype.time_unit  # type: ignore[attr-defined]
-                s_cast = (
-                    s.astype("Int64[pyarrow]") if is_pyarrow_dtype else s.astype("int64")
-                )
-                result = calculate_timestamp_datetime(
-                    s_cast, original_time_unit, time_unit
-                )
-            else:
-                msg = "Input should be either of Date or Datetime type"
-                raise TypeError(msg)
-            return result.where(~mask_na)
-
-        return self._compliant_expr._from_call(
-            func,
-            "datetime",
-            time_unit=time_unit,
-            returns_scalar=self._compliant_expr._returns_scalar,
-        )
-
-    def total_minutes(self) -> DaskExpr:
-        return self._compliant_expr._from_call(
-            lambda _input: _input.dt.total_seconds() // 60,
-            "total_minutes",
-            returns_scalar=self._compliant_expr._returns_scalar,
-        )
-
-    def total_seconds(self) -> DaskExpr:
-        return self._compliant_expr._from_call(
-            lambda _input: _input.dt.total_seconds() // 1,
-            "total_seconds",
-            returns_scalar=self._compliant_expr._returns_scalar,
-        )
-
-    def total_milliseconds(self) -> DaskExpr:
-        return self._compliant_expr._from_call(
-            lambda _input: _input.dt.total_seconds() * 1000 // 1,
-            "total_milliseconds",
-            returns_scalar=self._compliant_expr._returns_scalar,
-        )
-
-    def total_microseconds(self) -> DaskExpr:
-        return self._compliant_expr._from_call(
-            lambda _input: _input.dt.total_seconds() * 1_000_000 // 1,
-            "total_microseconds",
-            returns_scalar=self._compliant_expr._returns_scalar,
-        )
-
-    def total_nanoseconds(self) -> DaskExpr:
-        return self._compliant_expr._from_call(
-            lambda _input: _input.dt.total_seconds() * 1_000_000_000 // 1,
-            "total_nanoseconds",
-            returns_scalar=self._compliant_expr._returns_scalar,
-        )
-
-
-class DaskExprNameNamespace:
-    def __init__(self: Self, expr: DaskExpr) -> None:
-        self._compliant_expr = expr
-
-    def keep(self: Self) -> DaskExpr:
-        root_names = self._compliant_expr._root_names
-
-        if root_names is None:
-            msg = (
-                "Anonymous expressions are not supported in `.name.keep`.\n"
-                "Instead of `nw.all()`, try using a named expression, such as "
-                "`nw.col('a', 'b')`\n"
-            )
-            raise ValueError(msg)
-
-        return self._compliant_expr.__class__(
-            lambda df: [
-                series.rename(name)
-                for series, name in zip(self._compliant_expr._call(df), root_names)
-            ],
-            depth=self._compliant_expr._depth,
-            function_name=self._compliant_expr._function_name,
-            root_names=root_names,
-            output_names=root_names,
-            returns_scalar=self._compliant_expr._returns_scalar,
-            backend_version=self._compliant_expr._backend_version,
-            version=self._compliant_expr._version,
-            kwargs=self._compliant_expr._kwargs,
-        )
-
-    def map(self: Self, function: Callable[[str], str]) -> DaskExpr:
-        root_names = self._compliant_expr._root_names
-
-        if root_names is None:
-            msg = (
-                "Anonymous expressions are not supported in `.name.map`.\n"
-                "Instead of `nw.all()`, try using a named expression, such as "
-                "`nw.col('a', 'b')`\n"
-            )
-            raise ValueError(msg)
-
-        output_names = [function(str(name)) for name in root_names]
-
-        return self._compliant_expr.__class__(
-            lambda df: [
-                series.rename(name)
-                for series, name in zip(self._compliant_expr._call(df), output_names)
-            ],
-            depth=self._compliant_expr._depth,
-            function_name=self._compliant_expr._function_name,
-            root_names=root_names,
-            output_names=output_names,
-            returns_scalar=self._compliant_expr._returns_scalar,
-            backend_version=self._compliant_expr._backend_version,
-            version=self._compliant_expr._version,
-            kwargs={**self._compliant_expr._kwargs, "function": function},
-        )
-
-    def prefix(self: Self, prefix: str) -> DaskExpr:
-        root_names = self._compliant_expr._root_names
-        if root_names is None:
-            msg = (
-                "Anonymous expressions are not supported in `.name.prefix`.\n"
-                "Instead of `nw.all()`, try using a named expression, such as "
-                "`nw.col('a', 'b')`\n"
-            )
-            raise ValueError(msg)
-
-        output_names = [prefix + str(name) for name in root_names]
-        return self._compliant_expr.__class__(
-            lambda df: [
-                series.rename(name)
-                for series, name in zip(self._compliant_expr._call(df), output_names)
-            ],
-            depth=self._compliant_expr._depth,
-            function_name=self._compliant_expr._function_name,
-            root_names=root_names,
-            output_names=output_names,
-            returns_scalar=self._compliant_expr._returns_scalar,
-            backend_version=self._compliant_expr._backend_version,
-            version=self._compliant_expr._version,
-            kwargs={**self._compliant_expr._kwargs, "prefix": prefix},
-        )
-
-    def suffix(self: Self, suffix: str) -> DaskExpr:
-        root_names = self._compliant_expr._root_names
-        if root_names is None:
-            msg = (
-                "Anonymous expressions are not supported in `.name.suffix`.\n"
-                "Instead of `nw.all()`, try using a named expression, such as "
-                "`nw.col('a', 'b')`\n"
-            )
-            raise ValueError(msg)
-
-        output_names = [str(name) + suffix for name in root_names]
-
-        return self._compliant_expr.__class__(
-            lambda df: [
-                series.rename(name)
-                for series, name in zip(self._compliant_expr._call(df), output_names)
-            ],
-            depth=self._compliant_expr._depth,
-            function_name=self._compliant_expr._function_name,
-            root_names=root_names,
-            output_names=output_names,
-            returns_scalar=self._compliant_expr._returns_scalar,
-            backend_version=self._compliant_expr._backend_version,
-            version=self._compliant_expr._version,
-            kwargs={**self._compliant_expr._kwargs, "suffix": suffix},
-        )
-
-    def to_lowercase(self: Self) -> DaskExpr:
-        root_names = self._compliant_expr._root_names
-
-        if root_names is None:
-            msg = (
-                "Anonymous expressions are not supported in `.name.to_lowercase`.\n"
-                "Instead of `nw.all()`, try using a named expression, such as "
-                "`nw.col('a', 'b')`\n"
-            )
-            raise ValueError(msg)
-        output_names = [str(name).lower() for name in root_names]
-
-        return self._compliant_expr.__class__(
-            lambda df: [
-                series.rename(name)
-                for series, name in zip(self._compliant_expr._call(df), output_names)
-            ],
-            depth=self._compliant_expr._depth,
-            function_name=self._compliant_expr._function_name,
-            root_names=root_names,
-            output_names=output_names,
-            returns_scalar=self._compliant_expr._returns_scalar,
-            backend_version=self._compliant_expr._backend_version,
-            version=self._compliant_expr._version,
-            kwargs=self._compliant_expr._kwargs,
-        )
-
-    def to_uppercase(self: Self) -> DaskExpr:
-        root_names = self._compliant_expr._root_names
-
-        if root_names is None:
-            msg = (
-                "Anonymous expressions are not supported in `.name.to_uppercase`.\n"
-                "Instead of `nw.all()`, try using a named expression, such as "
-                "`nw.col('a', 'b')`\n"
-            )
-            raise ValueError(msg)
-        output_names = [str(name).upper() for name in root_names]
-
-        return self._compliant_expr.__class__(
-            lambda df: [
-                series.rename(name)
-                for series, name in zip(self._compliant_expr._call(df), output_names)
-            ],
-            depth=self._compliant_expr._depth,
-            function_name=self._compliant_expr._function_name,
-            root_names=root_names,
-            output_names=output_names,
-            returns_scalar=self._compliant_expr._returns_scalar,
-            backend_version=self._compliant_expr._backend_version,
-            version=self._compliant_expr._version,
-            kwargs=self._compliant_expr._kwargs,
-        )
+    @property
+    def name(self: Self) -> DaskExprNameNamespace:
+        return DaskExprNameNamespace(self)
